I have the basics of the fast spiking neuron working. It did slow down compared to my earlier tests, but not horribly so. It is about 3 times slower. I am doing a lot more global IO now though, so that it is understandable. One thing that still needs to be tested is how the speed will change when I have random number of synaptic connections instead of a uniform number that I am using now.

+++ trunk/Libraries/OpenNeuronCL/FastSpikingNeuralModel.h
@@ -18,7 +18,8 @@ namespace OpenNeuronCL
 			unsigned short m_iSynapsesPerNeuron;
 
 			vector<float> m_aryTestData1;
-			vector<unsigned int> m_aryTestData2;
+			vector<float> m_aryTestData2;
+			//vector<unsigned int> m_aryTestData2;
 
 			cl::NDRange m_FSN_Global;
 			cl::NDRange m_FSN_Local;
